+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## Unrleased
+
+* Added: Host entry commands
+
+
 ## 1.2.1 - 2017-07-19
 
 * Fixed: Wrong existence checks in Set-FileExplorerNamespace and Remove-FileExplorerNamespace